move discord context code to the discord interface module
