+++ xlive/Blam/Engine/networking/panorama/panorama_presence.cpp
@@ -1,7 +1,10 @@
 #include "stdafx.h"
 #include "panorama_presence.h"
 
-void __cdecl networking_panorama_presence_set_presence(e_context_presence context)
+/* public code */
+
+void __cdecl networking_panorama_presence_set_presence(uint32 context)
 {
 	INVOKE(0x1B07A0, 0x0, networking_panorama_presence_set_presence, context);
+	return;
 }
